# Make Math.Pow, Min and Max for doubles intrinsics