--- a/mono/mini/intrinsics.c
+++ b/mono/mini/intrinsics.c
@@ -132,12 +132,13 @@ llvm_emit_inst_for_method (MonoCompile *cfg, MonoMethod *cmethod, MonoMethodSign
 			ins->type = STACK_R8;
 			ins->dreg = mono_alloc_dreg (cfg, (MonoStackType)ins->type);
 			ins->sreg1 = args [0]->dreg;
-			if (fsig->param_count == 2) { // POW
+			if (fsig->param_count > 1) { // POW
 				ins->sreg2 = args [1]->dreg;
-			} else if (fsig->param_count == 3) { // FMA
-				ins->sreg2 = args [1]->dreg;
+			}
+			if (fsig->param_count > 2) { // FMA
 				ins->sreg3 = args [2]->dreg;
 			}
+			g_assert (fsig->param_count <= 3);
 			MONO_ADD_INS (cfg->cbb, ins);
 		}
 	}
@@ -176,13 +177,13 @@ llvm_emit_inst_for_method (MonoCompile *cfg, MonoMethod *cmethod, MonoMethodSign
 			ins->type = STACK_R8;
 			ins->dreg = mono_alloc_dreg (cfg, (MonoStackType)ins->type);
 			ins->sreg1 = args [0]->dreg;
-			if (fsig->param_count == 2) { // POW, MIN, MAX
+			if (fsig->param_count > 1) { // POW, MIN, MAX
 				ins->sreg2 = args [1]->dreg;
 			}
-			if (fsig->param_count == 3) { // FMA
-				ins->sreg2 = args [1]->dreg;
+			if (fsig->param_count > 2) { // FMA
 				ins->sreg3 = args [2]->dreg;
 			}
+			g_assert (fsig->param_count <= 3);
 			MONO_ADD_INS (cfg->cbb, ins);
 		}