--- a/mono/mini/llvm-jit.cpp
+++ b/mono/mini/llvm-jit.cpp
@@ -426,20 +426,20 @@ mono_llvm_create_ee (AllocCodeMemoryCb *alloc_cb, FunctionEmittedCb *emitted_cb,
 	EnableMonoEH = true;
 	MonoEHFrameSymbol = "mono_eh_frame";
 
-	TargetOptions opts;
 	if (mono_use_fast_math) {
+		TargetOptions opts;
 		opts.NoInfsFPMath = true;
 		opts.NoNaNsFPMath = true;
 		opts.NoSignedZerosFPMath = true;
 		opts.NoTrappingFPMath = true;
 		opts.UnsafeFPMath = true;
 		opts.AllowFPOpFusion = FPOpFusion::Fast;
+		EB.setTargetOptions (opts);
 	}
 
 	EngineBuilder EB;
 	EB.setOptLevel(CodeGenOpt::Aggressive);
 	EB.setMCPU(sys::getHostCPUName());
-	EB.setTargetOptions (opts);
 	auto TM = EB.selectTarget ();
 	assert (TM);