Before this PR `Pow`, `Min` and `Max` worked only for `float`

```csharp
[MethodImpl(MethodImplOptions.NoInlining)]
static float Test1(float x) => MathF.Pow(x, 2);

[MethodImpl(MethodImplOptions.NoInlining)]
static double Test2(double x) => Math.Pow(x, 2);

// Min and Max require `--ffast-math` mode for optimal codegen

[MethodImpl(MethodImplOptions.NoInlining)]
static float Test3(float x, float y) => MathF.Min(x, y);

[MethodImpl(MethodImplOptions.NoInlining)]
static double Test4(double x, double y) => Math.Min(x, y);

[MethodImpl(MethodImplOptions.NoInlining)]
static float Test5(float x, float y) => Math.Max(x, y);

[MethodImpl(MethodImplOptions.NoInlining)]
static double Test6(double x, double y) => Math.Max(x, y);
```

Asm:
```asm

0000000000000000	vmulss	%xmm0, %xmm0, %xmm0
0000000000000004	retq

0000000000000000	vmulsd	%xmm0, %xmm0, %xmm0
0000000000000004	retq

0000000000000000	vminss	%xmm0, %xmm1, %xmm0
0000000000000004	retq

0000000000000000	vminsd	%xmm0, %xmm1, %xmm0
0000000000000004	retq

0000000000000000	vmaxss	%xmm0, %xmm1, %xmm0
0000000000000004	retq

0000000000000000	vmaxsd	%xmm0, %xmm1, %xmm0
0000000000000004	retq
```

LLVM IR: (Optimized, after InstCombine, etc)
```llvm
define monocc float @"HelloWorld.Program:Test1 (single)"(float %arg_x) #0 {
BB0:
  %pow2 = fmul float %arg_x, %arg_x
  ret float %pow2
}

define monocc double @"HelloWorld.Program:Test2 (double)"(double %arg_x) #0 {
BB0:
  %pow2 = fmul double %arg_x, %arg_x
  ret double %pow2
}

define monocc float @"HelloWorld.Program:Test3 (single,single)"(float %arg_x, float %arg_y) #0 {
BB0:
  %.inv = fcmp ogt float %arg_x, %arg_y
  %t26.v = select i1 %.inv, float %arg_y, float %arg_x
  ret float %t26.v
}

define monocc double @"HelloWorld.Program:Test4 (double,double)"(double %arg_x, double %arg_y) #0 {
BB0:
  %.inv = fcmp ogt double %arg_x, %arg_y
  %t21 = select i1 %.inv, double %arg_y, double %arg_x
  ret double %t21
}

define monocc float @"HelloWorld.Program:Test5 (single,single)"(float %arg_x, float %arg_y) #0 {
BB0:
  %.inv = fcmp olt float %arg_x, %arg_y
  %t21.v = select i1 %.inv, float %arg_y, float %arg_x
  ret float %t21.v
}

define monocc float @"HelloWorld.Program:Test5 (single,single)"(float %arg_x, float %arg_y) #0 {
BB0:
  %.inv = fcmp olt float %arg_x, %arg_y
  %t21.v = select i1 %.inv, float %arg_y, float %arg_x
  ret float %t21.v
}
```
CoreCLR: [sharplab.io](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABAJgEYBYAKGIAYACY8gOgCUBXAOwwEt8YLAMIR8AB14AbGFADKMgG68wMXAG4axAMxNSDIQwDeNGgzMMA2gFkYGABYQAJgElxkgBQ37T12MkB5MT4ILlwWADkIZy5JXi44gHMASgBdU3NmJAYAM0kIbAwGABVVDHJ3XPzChCSGAF4APgYrArsAMRYABQgAd3cENAZSJI1qdLNrWwcXN08pnzdA4NCIqJi4xNTxpnIsxwgOYGli0tJ3fcPjmvqmlvsu3v7B4dHtye8ZvzmP3wCg3hCYUi0Vi8S4yTS1HMOyylQKJ1wGC0FTy8IGOVRhQAnrVGs1Wh0rHEngwca8oeZ3tNft9qYt/oDViCNuCthSzJkGBcjjAERgUOcDjyGOjucccTd8fciVwSWSTOzLF46V9lQs/EsAStgeswRDtpy4YUSoiAKwoqoiwZG0m426tFgtBBykYK6FU9UeNWfP7LIFrUGbSHQzli3kmjBIQWXXmioXiu1SuyO7DO9HyxU0AC+QA==) (it doesn't unroll Math.Pow and doesn't have fast math mode for vmax)