@filipnavara 2019-08-29
## Burgers benchmark

### Before

|    Method |       Mean |   Error |  StdDev |     Median |        Min |        Max |      Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|---------- |-----------:|--------:|--------:|-----------:|-----------:|-----------:|-----------:|-----------:|-----------:|----------:|
| Burgers_0 | 2,665.4 ms | 5.39 ms | 4.21 ms | 2,665.7 ms | 2,654.4 ms | 2,669.3 ms | 46000.0000 | 46000.0000 | 46000.0000 |         - |
| Burgers_1 | 2,251.6 ms | 5.40 ms | 4.79 ms | 2,250.4 ms | 2,245.7 ms | 2,264.1 ms |          - |          - |          - |         - |
| Burgers_2 |   149.9 ms | 0.52 ms | 0.48 ms |   149.7 ms |   149.3 ms |   150.6 ms |          - |          - |          - |         - |
| Burgers_3 | 2,414.9 ms | 5.80 ms | 5.14 ms | 2,412.9 ms | 2,411.1 ms | 2,429.8 ms |          - |          - |          - |         - |

### After

|    Method |       Mean |   Error |  StdDev |     Median |        Min |        Max |      Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|---------- |-----------:|--------:|--------:|-----------:|-----------:|-----------:|-----------:|-----------:|-----------:|----------:|
| Burgers_0 |   537.3 ms | 7.10 ms | 6.64 ms |   537.1 ms |   523.8 ms |   546.2 ms | 45000.0000 | 45000.0000 | 45000.0000 |         - |
| Burgers_1 |   151.6 ms | 1.63 ms | 1.52 ms |   152.2 ms |   149.4 ms |   154.0 ms |          - |          - |          - |         - |
| Burgers_2 |   150.0 ms | 0.45 ms | 0.42 ms |   150.1 ms |   149.3 ms |   150.7 ms |          - |          - |          - |         - |
| Burgers_3 | 2,418.1 ms | 5.25 ms | 4.65 ms | 2,417.6 ms | 2,412.7 ms | 2,430.5 ms |          - |          - |          - |         - |