@EgorBo 2019-08-29
@monojenkins build Windows x64 FullAOT