@tannergooding 2019-08-29
`Math.Max` to `vmaxss` is not an equivalent translation. They handle NaN and +/-0 differently, so it is not safe to do unless the user has opted into some kind of "fast math" mode.