@lewing 2019-08-29
@tannergooding that is what `--ffast-math` does see the `mono_use_fast_math` checks in the changes.